Rewrite the mailcap support
* configure.ac: Build libmailutils/mailcap/Makefile

* include/mailutils/mailcap.h: Rewrite
* include/mailutils/sys/Makefile.am: Add mailcap.h
* include/mailutils/sys/mailcap.h: New file.
* include/mailutils/util.h: Add missing include.
* lib/mailcap.c: Rewrite using new API.
* libmailutils/Makefile.am: Build in mailcap
* libmailutils/base/Makefile.am: Remove mailcap.c
* libmailutils/base/mailcap.c: Remove.

* libmailutils/mailcap/Makefile.am: New file.
* libmailutils/mailcap/ctmatch.c: New file.
* libmailutils/mailcap/deferrclos.c: New file.
* libmailutils/mailcap/mcp_creat.c: New file.
* libmailutils/mailcap/mcp_destr.c: New file.
* libmailutils/mailcap/mcp_err.c: New file.
* libmailutils/mailcap/mcp_flags.c: New file.
* libmailutils/mailcap/mcp_get.c: New file.
* libmailutils/mailcap/mcp_sel.c: New file.
* libmailutils/mailcap/mcp_count.c: New file.
* libmailutils/mailcap/mcp_find.c: New file.
* libmailutils/mailcap/ent_cmd.c: New file.
* libmailutils/mailcap/ent_creat.c: New file.
* libmailutils/mailcap/ent_destr.c: New file.
* libmailutils/mailcap/ent_locus.c: New file.
* libmailutils/mailcap/ent_type.c: New file.
* libmailutils/mailcap/fieldacc.c: New file.
* libmailutils/mailcap/fields.c: New file.
* libmailutils/mailcap/finder.c: New file.
* libmailutils/mailcap/parse.c: New file.
* libmailutils/mailcap/parsefile.c: New file.

* libmailutils/tests/mailcap.c: Rewrite using new API. Improve cli
* libmailutils/tests/mcf.c: New file.
* libmailutils/tests/ctm.c: New file.
* libmailutils/tests/ctm.at: New file.
* libmu_cpp/mailcap.cc: Rewrite using new API.
* python/2/libmu_py/mailcap.c: Rewrite using new API.
* python/3/libmu_py/mailcap.c: Rewrite using new API.
* examples/python/2/mailcap.py: Minor change. This example is
broken with the new API. Should be fixed.
* include/mailutils/cpp/mailcap.h: Update
* include/mailutils/cpp/pop3.h: Update

+++ include/mailutils/cpp/pop3.h
@@ -56,7 +56,7 @@ class Pop3
   void quit ();
   Stream& retr (unsigned int msgno);
   void rset ();
-  void stat (unsigned int* count, mu_off_t* octets);
+  void stat (size_t* count, mu_off_t* octets);
   Stream& top (unsigned int msgno, unsigned int lines);
   std::string uidl  (unsigned int msgno);
   Iterator& uidl_all ();
